:truck: rename: Validator 클래스 Model Directory로 이동
